Moved out Hackney Header & Logo onto the Layout root.
It's because I'm going to be using this header for both layouts after all. I need pretty much everything about it for both layouts. The only difference I'll MAYBE need is for the Hackney Logo to be pushed to the left for Docs pages. Will figure how to do it when I need it, if I need it.
