+++ components/Layout/Docs/index.js
@@ -1,4 +1,4 @@
-import Header from "./Header";
+import Header from "../Header";
 
 const DocsLayout = ({ children }) => (
   <>
