+++ components/Layout/Non-Docs/index.js
@@ -1,4 +1,4 @@
-import Header from "./Header";
+import Header from "../Header";
 
 const NonDocsLayout = ({ children }) => (
   <>
